terminando sugerencias en el back

+++ frontend/main/forms/food_forms.py
@@ -1,6 +1,6 @@
 # - *- coding: utf- 8 - *-
 from flask_wtf import FlaskForm #Importa funciones de formulario
-from wtforms import PasswordField, SubmitField, StringField, SelectField, HiddenField, IntegerField  # Importa campos
+from wtforms import PasswordField, SubmitField, StringField, SelectField, HiddenField, IntegerField, FloatField  # Importa campos
 from wtforms.fields import EmailField,DateField, TimeField #Importa campos HTML
 from wtforms import validators #Importa validaciones
 from flask_login import current_user
@@ -9,7 +9,7 @@ class FoodForm(FlaskForm):
 
     name = StringField('Introduzca el nombre del alimento:')
 
-    carbohydrates = IntegerField('Introduzca la cantidad de carbohidratos que posee el alimento:')
+    carbohydrates = FloatField('Introduzca la cantidad de carbohidratos que posee el alimento:')
